src/tests/functions_test.py
import unittest

from src.functions import markdown_to_blocks, markdown_to_html_node, text_to_textnodes
from src.textnode import TextNode

text_type_text = "text"
text_type_bold = "bold"
text_type_italic = "italic"
text_type_code = "code"
text_type_image = "image"
text_type_link = "link"


class TextToTextNodesTestCase(unittest.TestCase):
    def test_text_seq_1(self):
        text = "This is **text** with an *italic* word and a `code block` and an ![obi wan image](https://i.imgur.com/fJRm4Vk.jpeg) and a [link](https://boot.dev)"

        expected_nodes = [
            TextNode("This is ", text_type_text),
            TextNode("text", text_type_bold),
            TextNode(" with an ", text_type_text),
            TextNode("italic", text_type_italic),
            TextNode(" word and a ", text_type_text),
            TextNode("code block", text_type_code),
            TextNode(" and an ", text_type_text),
            TextNode(
                "obi wan image", text_type_image, "https://i.imgur.com/fJRm4Vk.jpeg"
            ),
            TextNode(" and a ", text_type_text),
            TextNode("link", text_type_link, "https://boot.dev"),
        ]

        self.assertEqual(text_to_textnodes(text), expected_nodes)


class MarkdownToBlockTestCase(unittest.TestCase):
    def test_text_seq_1(self):
        markdown = "# This is a heading\n\n This is a paragraph of text. It has some **bold** and *italic* words inside of it.\n\n* This is the first list item in a list block\n* This is a list item\n* This is another list item"

        # Output should contain blocks seperated by \n\n
        expected_blocks = [
            "# This is a heading",
            "This is a paragraph of text. It has some **bold** and *italic* words inside of it.",
            "* This is the first list item in a list block\n* This is a list item\n* This is another list item",
        ]
        self.assertEqual(markdown_to_blocks(markdown), expected_blocks)


class MarkdownToHTMLNodeTestCase(unittest.TestCase):
    def test_text_seq_1(self):
        markdown = "# This is a h1 tag\n\n This is a paragraph of text. It has some **bold** and *italic* words inside of it.\n\n* This is the first list item in a list block\n* This is a list item\n* This is another list item"

        # Output should contain blocks seperated by \n\n
        expected_html = "<div id='root'><h1>This is a h1 tag</h1><p>This is a paragraph of text. It has some <b>bold</b> and <i>italic</i> words inside of it.</p><ul><li>This is the first list item in a list block</li><li>This is a list item</li><li>This is another list item</li></ul></div>"

        self.assertEqual(markdown_to_html_node(markdown), expected_html)
